add Romo.setup API for initializing and setting up the Romo runtime
This sets up Romo for use in the browser. When the browser is
ready, it sets up the initial state. To get this working, I had
to add API for:

* Array conversion
* Function/element binding
* Event handling

In future efforts, I'll fill out the rest of the Romo API and
bring in the planned Utilities and Components.