add Romo.XHR DOM component
This component binds to a DOM element and adds XHR behaviors to it.
You can configure the XHR behaviors using DOM data attributes and
you can interact with it using DOM events.

This also fixes some bugs in the XMLHttpRequest utility that I
found while testing out this new component.