switch to using Lighttpd web server to host tests
This adds a few bins and some example configuration for setting
up running the tests. I had to use a proper webserver since CORS
started blocking embedded dependencies I was trying to load from
Skypack.dev.

This is prep for adding the Romo.Spinner DOM component as it
requires spinjs which we bring in via Skypack.

I chose Lighttpd since I could install it with brew and run it
on a local document root without doing any extra configuration
outside of the repo.