rework Romo.DOM enumerator methods to emit DOM objects by default
This switches the default Array/Enumerator type methods to emit
Romo.DOM instances instead of the raw Elements by default. This
follows the principle of least surprise in that if you are
operating on DOM, you expect to get DOM instance to operate on.

This also adds *Element equivalent methods that continue emitting
Elements and backports the existing usages to call those since they
were already expecting/operating on Elements.