rework smoke tests and bin
This switches both the system tests and the smoke test to import
all dependencies to ensure everything gets loaded in the correct
order. This also adds comments for smoke test equivalents in the
system tests for reference.

This also updates the smoke_test bin to take a simpler, less
automated approach.