add separate and additional publish for Romo UI
Having a separate package makes it nice for explicitly choosing
whether to bring in the UI extras or not. This enables using
e.g. Skypack.dev. However I didn't want the overhead of having
two distinct repos when one hard-depends on another. I also
wasn't wild about versioning them independently b/c confusion.

This adds an additional package.json and package for Romo UI and
then creates a publish bin for hijacking the default publish
using the UI's package JSON. This allows me to develop, version,
and release the packages in sync and ensures that changes from
one package don't break another.

Note: I maybe could have done this by using a monorepo type
pattern but this was easier to proof-of-concept. I may move
to a monorepo in the future when things stabalize - not sure.