add Romo.UI.PopoverStack
This custom utility class models a stack of Romo.Popover component
instances. Since popovers naturally stack on top of the page and
each other, it is natural to manage them with a stack-like object.

This also sets up the global `Romo.ui` that holds the popover
stack selectors state and updates Romo UI to initialize it
properly.