don't default Romo.XHR responseTypes to 'json'
This was an invalid assumption and was too presumptuous. A better
default is to not specify any response type and let the response
come back as text. Then if the user wants decoded output, e.g.
JSON, they can specify that.

This also reorgs the XHR tests to list the most relevent data
attributes first for readability.