update Romo.Form: don't default JSON responseType, don't default redirect GET submissions
This updates the Form components to not default a JSON response
type if none is specified. This is too specific of a default and
can cause surprises. A better default is to specify no response
type and return text.

This also updates the component to not default GET XHR submissions
to redirect the page. Again, this is too specific for default
behavior of an XHR Form. A better default is to just emit the
response data in the success event like Romo.XHR does. The user
can turn on the page redirection if desired.

I noticed these b/c I got surprised by them when testing Forms
embedded in the coming Romo.UI.Frame component.