update Romo.Spinner to optionally start/stop immediately
By default, the spinner starts/stops in a `pushFn` to ensure all
other event handling is processed before starting/stopping the
Spinner. This ensures all UI mutations are done so the spinner
can lock them in when it starts. This also about being as
late-bound as possible.

However, there are times when it is necessary to _immediately_
start/stop the spinners to coordinate with other synchronous
behavior. This is usually only necessary when another component
is composing Spinner behavior into itself and needs to manage the
Spinner tightly. This updates the start/stop API to take an
optional `{start|stop}Immediately` control param to force this.

This is specifically needed for the new Romo.UI.Frame component
so it can orchestrate its use of an internal Spinner.