update Romo.XHR to handle disabledWithSpinner better
This comes after testing/using disableWithSpinner on the Tooltip
XHR integration. This stops the spinner immediately on enable to
allow immediately updating the content of the XHR dom element.
This also moves the disableWithSpinner state handling to inside
the `_startCall` method to ensure it happens JIT and for
readability in the dequeueCall method.

This also adds a zero-ms delay to kicking off the XHR. This just
makes debugging easier if you want to artificially insert XHR
delay. In the future we may add a DOM configuration setting to
allow ad-hoc XHR min-delays.

Finally, this adds emitting the component instance on the
Romo.XHR:callEnd event. This should have been there originally
per convention but was missed.