update Romo.UI.Tooltip to work with XHR disable with spinner
This comes after additional testing where I inserted artificial
XHR delay but wasn't seeing the spinner. This reworks the place
functions and update content functions to allow for synchronous
callbacks if desired. Then the XHR integration uses callbacks to
ensure the spinner is allowed to start/stop before the tooltip
content is updated so that it is actually shown.

This also updates the CSS to specify a min-width/height of 1rem
so the spinner has a somewhat appropriate size.