add Romo.Location and compose in Romo.PageState
This extracts Location handling since the native `Location` object
isn't very friendly and I want to commonize URL and pathname/search
handling. This also composes this new object in Romo.PageState so
they have a similar API.

A side-effect of this is some new features in Romo.PageState:
in addition to the API methods for mutating the search params,
you can mutate the pathname, the search string, or the location
(as a macro for both pathname/search mutation) using setters. This
gives you more options for changing the page state.