update Romo.updateHTML to fallback to Romo.updateText as needed
It became annoying to call `Romo.updateHTML` and then in some
scenarios the "HTML" would just be a string of text. This makes the
updateHTML API method a bit friendlier in that it will fallback
to updateText if there are no HTML elements in the given HTML
string.

I noticed this while building out Romo.UI.LocalTime's integration
with Romo.UI.Tooltip.