add DOM data handling methods to Romo.DOMComponent
This updates Romo.DOMComponent to optionally take a domDataName
attribute and, if given, provide utility methods for interacting
with interacting with DOM data attributes without having to
repeat the DOM data name.

This came up when building out Romo.UI.LocalTime. It has many
supported DOM data attributes and it became a pain and was noisy
repeating `romo-ui-local-time` over and over again.

In a future effort, I'll go and back-port this API to the existing
component implementations.