reorg the components and utilities
This updates it so that anything classified as "api" or "utility"
goes into Romo and anything classified as a "component" goes into
Romo.UI. This also removes a lot of the intermediate directories
and import js files.

There are two goals here that became apparent after adding a bunch
of components:

* the organization was too complex and involved which - this aims
  to simplify things a bit.
* Romo base should include everything needed for a user to build
  their _own_ "Romo UI". Everything in Romo.UI is just an
  implementation of UI components built on the API/Utilities
  provided by Romo base.

Note: I still need to rename the Form, Spinner, and XHR components
in the UI namespace but I'm going to do that in a follow up effort
after I backport the data handling API in Romo.DOMComponent. This
will make the port a lot easier.