remove reloadPage handling from Romo.XHR and Romo.Form
This is prep for adding API/support for handling banner/toast
Alerts. The reload page logic was tied up with the alert logic.
Since I am about to rework the alert logic, I had to deal with it.
Upon more reflection, I'm uncomfortable with the alert logic and
I feel it binds the user and isn't flexble/customizable enough.
I feel the same about the reload logic. There are a number of
cases where you want to show an alert and after the users cofirms
the alert, then you want to reload the page. This current logic
doesn't support that well and that is also a highly specific type
of workflow.

Given all this, I'm choosing to remove this from the XHR and Form
APIs in favor of the user configuring or more manually coding up
their desired reload workflows.