remove the Romo.alert/FlashAlert handling; replace with Banner/Toast alert handling
This is more generic in that it just defines two types of alerts
in terms of presentation. By default both alerts just log errors
on the console but this configuration can be overridden and
extended. XmlHttpRequest (and therefore XHR and Form) supports
detecting banner/toast alerts in JSON responses.

This is prep for adding Banner/Toast UI components that handle
presenting these two alert types.