add Romo.UI.Banner
This shows alert-intended markup in a non-stack-popover that can
only be closed if a `[data-romo-ui-banner-close]` child element is
clicked OR by executing `Romo.UI.Banner.doHide`.

This is a singleton alert - only 1 set of banner markup can be
shown at a time.

The tests page demonstrates how any banners in auto-init markup
are auto-shown. It also demonstrates how banner content embedded in
XHR/Form JSON responses are also auto-shown.