cleanup XmlHTTPRequest error and banner/toast in JSON response handling
I was a bit too aggressive in the removals I did in PR 27. I was
trying to simplify the error handling in Romo.xhr responses and
accidentally removed the handling and configuration for non-muted
error responses. This restores the basic config/handling just
without any of the reload page stuff.

This also renames the banner/toasts handling to be explicit about
it only applying to JSON response types. This is only needed for
JSON response types as other response types either can't support
embedded banners/toasts or e.g. HTML responses just embed the
banner/toast directly in their markup.