remove specialized Form submit spinner handling, fix a bug
I noticed this while investigating the bug fix of a typo in this
commit. We don't need to support this super esoteric method of
finding submit spinners in the interest of keeping things
simply. This is a relic of a previous implementation/take on this
I did in a different project.