add Romo.UI.Avatar
This component manages an avatar with support for fallbacks when
images fail to load. It starts by trying to load an image with
the configured `url` data attr. If that fails to load, it tries
to load an image with the configured `fallback-url` data attr. If
that fails to load, it falls back to inserting the configured
`abbrev` data attribute or finally falling back to abbreviating the
configured `name` data attr.

This gives lots of options for representing a named person, group,
or thing with late-bound fallbacks to handle a variety of
scenarios.