update Romo.XHRCache to read from dynamic URL
Previously, the XHR cache was bound to a static URL. This meant if
the URL you wanted to read/cache changed, you had re-initialize a
new XHR Cache. This wasn't the friendliest API ever and could lead
to surprising behavior if you didn't realize the cache was bound
to the URL.

This switches to making the URL be passed in on read calls. This
means the cache is no longer bound to a single URL and can be
reused in a late-bound way on dynamic URLs.

This also fixes a bug where the content type wasn't getting set
properly on the cached responses.

Finally, this adds some system tests to ensure the utility works
as expected.