update Romo.UI.Dropdown placement logic to fix some bugs
Previously, the dropdown positioned itself using only the top
and left CSS attributes. To calculate the left when positioning
the dropdown on the right side of its anchor, you had to use the
popover width in the calculation. Similarly, to calculate the top
when positioning the dropdown on the top side of its anchor, you
had to use the popover height in the calculation.

Having to rely on the popover width/height in the calculations
is problematic as, in the absence of constraints, the popover
will auto resize itself as its position is altered. This means a
you have a "moving target" for your calculations and placement
bugs are inevitable.

This switches the algorithm to not rely on the width/heigh and
instead place using all 4 position options: top, right, bottom,
and left. This removes the "moving target" and therefore removed
some placement bugs.