add configuration for Romo.UI.Avatar default URL
This allows the user to configure their own global default avatar
URL. This also seeds the configuration with a hopefully unique
`/romo-ui/img/default-avatar` URL that:

* ...is absolute. Previously the undefined URL was relative so you
  risked accidentally matching on wildcard paths in the app.
* ...intention revealing. The URL hints at what it is looking for.

Finally this tweaks the main URL to use the fallback URL if no
value is specified. The fallback URL uses the configured default
URL if no value is specified. This is just to try and ensure there
is a value for the attributes and setup the precedence of those
values.