update Romo.UI.Dropdown to support ad-hoc open CSS classes
These CSS classes are set when the dropdown is open and removed
when the dropdown is closed. Use this to set custom styles to
help indicate an active dropdown is open.

This also adds the ability to configure the base open CSS value.
By default it is 'active'. This base CSS class is always added
when the dropdown is open, in addition to any ad-hoc CSS classes.