update Romo.UI.IndicatorTextInput to configure its display style
This came up when incorporating this component into the coming
SelectDropdown component. I wanted to specifically configure
block display on the component in this context. Since there is a
wrapper, you have to set the display style on the wrapper.

In addition, this also updates the default wrapper display style
to follow that of the input. This is a good default if no
specific display value has been speifically configured.