fix a popover sizing bug related to auto-detected max height
If the popover is configured to detect and set a max height, that
max height would "seed" the popover placement algorithm on
subsequent popover placements. This caused the popover to place
incorrectly as the algorithm expects the popover to be in an
unconstrained state size-wise.

This adds an api to unset the max height and calls that before
calculating the new max height to achieve that "blank slate"
state.