update popovers to have 'element' settings for their width and min-width
Previously an configured width value of 'element' would detect the
width of the toggle DOM and set the max-width to that. Why should
max width get all the love? This reworks it to have that 'element'
width option for all 3 widths: min/max and width.

This is a more straighforward, less surprising API and has the
same effect as the previous API.