add doUpdatePopoverBodyDOM Popover API
This ensures you can update the popover content via a DOM object
in addition to the raw HTML API. This is handy when you have DOM
objects. Specifically I want to use this API in the coming
SelectDropdown component.

I also backport the DatePicker to use this API. Previously it was
manually updating the popover content which worked but bypassed
some logic like auto-placing the popover after the content had
been updated.