add Romo.UI.SelectDropdown component mixin
This composes IndicatorTextInput, OnKey, and Dropdown to build
a dropdown for use in selection UI e.g. a select input or other.

These dropdowns include a filter input at the top for filtering
large lists of select options. This also includes the base styling
and input/focus handling for interacting with the dropdown.

In a future effort, I'll use this component to compose and create
a select input component.