add keyboard navigation to Romo.UI.DatePicker
This is inspired by the keyboard navigation used in the Select
Dropdown component. The DatePicker is basically a select dropdown
bound to an indicator text input that uses a calendar layout for
the selection UI. Therefore it is reasonable to expect to use the
keyboard to highlight and select date values like you would in
a select dropdown.

This adds the behavior and limits the keyboard navigation to only
be in effect when the calendar dropdown is open.