add `data-romo-ui-select-dropdown-show-filter="false"` option
This allows configuring select dropdowns to not use a filter
input in the dropdown markup. This is just extra flexibility
for consumers of the component.

Note: I had to switch to binding on the window for the keydown
events for navigating options. This is b/c when the filter input
is hidden it can't receive key events. This is good b/c it
implicitly disables filtering the options. But it was bad b/c the
keydown events it normally received would propagate to the
container DOM and stopped propagating.