actually update SelectDropdown options when the dropdown is open
Previously, the only way to update the option list UI was to
close and reopen the dropdown. This is b/c we only updated the
options when the dropdown was opened. If is was already opened,
we just stored the new options but didn't try and update the list
UI.

This switches to, in addition to storing the new options, if the
dropdown is open also updating the list UI.

I didn't test thoroughly enough on previous iterations to this
logic. This adds a more thorough "input picker" test case to the
tests as well.

While working this, I noticed a bug in the `isPopover{Open|Closed}`
functions that I fix here too.