fix a bug with array Romo.URLSearchParams array value handling; add Romo.urlSearchURL
There was a bug un the `.toURL` method of `Romo.URLSearchParams`
where, because we were using `.get()`, array params would only
ever return the first value in array params. This switches to
using `.forEach` instead of a for-loop on `.keys()` with `.get()`.
This is a more robust method for getting each key/value pair when
processing the search params.

This also adds a `Romo.urlSearchURL` API method to complement
`Romo.urlSearch`. This is mostly used to exercise the `.toURL`
method on `Romo.URLSearchParams` but is also available now for
convenience if needed.