fix a bug in XHR POST Romo.UI.Form submissions
I had never formally tested this type of submission so when I
went to use this component and do these types of submissions in
an app, I noticed this bug. Now that we have formal testing in
place for this type of submission, I can repeat that bug.

This fixes the bug. There were two problems:

* I messed up the `Romo.XMLHttpRequest.Data.formData` getter and
  forgot to `return` the `formData` value. This meant it was
  always returning `undefined` and never sending any form data.
  This only occurred on XHR Posts that attempted to send form data.
* On `Romo.UI.Form.XHRPostSubmission`, I was converting the form
  values to a form data object and passing that as data to the
  underlying `Romo.xhr` call. However, that API expects primitive
  Object data. This switched to passing that instead.