fix Romo.UI.Form.BaseXHRSubmission error message handling
I didn't read the XHR spec carefully enough. It has
`responseText` and `responseXML` methods, so I guess I kinda
assumed it also had a `responseJSON` method. Well that was wrong.

For JSON responses, the `response` method will contain the JSON
if `responseType === 'json'`. This switches the handler logic to
conform to that spec.

Also, I didn't have a formal test for this so I added one.