fallback to status text error response text when XHR error response is null
This ensures we always return at least some form of error response
text when there is an XHR error. `null` isn't helpful at all.

This has mostly come up when using JSON response types and getting
a 500 on the server. In this case the alert will now say
"500 Internal Server Error" instead of "null" in the alert debug
message.